+++ cmd/basecli/commands/query.go
@@ -1,9 +1,11 @@
 package commands
 
 import (
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
 	wire "github.com/tendermint/go-wire"
+	lc "github.com/tendermint/light-client"
 	proofcmd "github.com/tendermint/light-client/commands/proofs"
 	"github.com/tendermint/light-client/proofs"
 
@@ -25,7 +27,9 @@ func doAccountQuery(cmd *cobra.Command, args []string) error {
 
 	acc := new(btypes.Account)
 	proof, err := proofcmd.GetAndParseAppProof(key, &acc)
-	if err != nil {
+	if lc.IsNoDataErr(err) {
+		return errors.Errorf("Account bytes are empty for address %X ", addr)
+	} else if err != nil {
 		return err
 	}